refactor: Move DeleteModal to components/modal
