+++ src/components/modal/DeleteModal.less
@@ -1,0 +1,41 @@
+@import '../../variables.less';
+
+.ant-modal-body {
+  padding: 0 24px;
+}
+
+.ant-modal-content {
+  border-radius: 20px;
+
+  .ant-modal-header {
+    border-bottom: 0px;
+    border-radius: 20px 20px 0 0;
+    padding: 1.5rem;
+  }
+
+  .ant-modal-footer {
+    border-top: 0px;
+    border-radius: 0 0 20px 20px;
+    padding: 1.5rem;
+
+    .ant-btn {
+      border-radius: 20px;
+    }
+  }
+
+  .error-text {
+    color: @red;
+    margin: 10px 0 0 0;
+  }
+
+  .icon-exclamation {
+    color: @red;
+    font-size: 1.25rem;
+    margin-bottom: 3px;
+    padding-right: 10px;
+  }
+
+  .modal-title > h3 {
+    display: inline;
+  }
+}
