--- src/components/my-promotions/DeleteModal.css
@@ -1,47 +1,0 @@
-/* Description text */
-.ant-modal-body {
-  padding: 0 24px;
-}
-
-/* Modal container */
-.ant-modal-content {
-  border-radius: 20px;
-}
-
-/* Footer button */
-.ant-modal-footer .ant-btn {
-  border-radius: 20px;
-}
-
-/* Error text */
-.error-text {
-  color: var(--red);
-  margin: 10px 0 0 0;
-}
-
-/* Title icon */
-.icon-exclamation {
-  color: var(--red);
-  font-size: 1.25rem;
-  margin-bottom: 3px;
-  padding-right: 10px;
-}
-
-/* Header */
-.modal-border-round .ant-modal-header {
-  border-bottom: 0px;
-  border-radius: 20px 20px 0 0;
-  padding: 1.5rem;
-}
-
-/* Footer */
-.modal-border-round .ant-modal-footer {
-  border-top: 0px;
-  border-radius: 0 0 20px 20px;
-  padding: 1.5rem;
-}
-
-/* Title */
-.modal-title > h3 {
-  display: inline;
-}
