+++ src/variables.less
@@ -1,3 +1,4 @@
 @black: #2e3a59;
 @orange: #ffa200;
 @yellow: #ffc529;
+@red: #f7263f;
